+++ CHANGES.rst
@@ -9,6 +9,9 @@ dev (master)
 * Accept ``iPAddress`` subject alternative name fields in TLS certificates.
   (Issue #258)
 
+* Fixed consistency of ``HTTPResponse.closed`` between Python 2 and 3.
+  (Issue #977)
+
 * ... [Short description of non-trivial change.] (Issue #)